requirements.backend.txt
fastapi==0.115.11
uvicorn==0.34.0
gunicorn==23.0.0
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
pydantic==2.10.6